requirements.txt
# Core CrewAI and LangChain
crewai>=0.28.0
langchain>=0.1.0
langchain-core>=0.1.0
langchain-community>=0.0.38

# LLM Providers
langchain-ollama>=0.1.0
langchain-groq>=0.1.0

# YouTube processing
youtube-transcript-api>=0.6.0

# Core utilities
python-dotenv>=1.0.0
requests>=2.31.0
beautifulsoup4>=4.12.0

# Data handling
pydantic>=2.0.0,<3.0.0

# Console and UI
rich>=13.0.0
colorama>=0.4.6

# Optional: Ollama client
ollama>=0.1.0
